test(administration): split the check spec into scenario files
check.spec.ts is close to the 500-line ceiling (ADR 2026-05-06) and the
findings-baseline work adds substantially more cases. Split it into a
check.spec/ directory by behavior group — units, checkExtensions
integration, and the runCheckCli entry point — mirroring the existing
setup.spec/ and report.spec/ layout. Pure move, no behavior change.
